Realizando testes no initenv.sh antes de realizar cópias

+++ initenv.sh
@@ -1,10 +1,18 @@
 #!/bin/bash
 
 # copiando diretório custom
-cp -a custom django/backup_corporativo/bkp
+if [ ! -d django/backup_corporativo/bkp/custom ];
+then
+    echo "copiando diretório custom";
+    cp -a custom django/backup_corporativo/bkp;
+fi
 
 # copiando setting.py
-cp -a django/backup_corporativo/settings_sample.py django/backup_corporativo/settings.py
+if [ ! -f django/backup_corporativo/settings.py ];
+then
+    echo "copiando settings.py"
+    cp -a django/backup_corporativo/settings_sample.py django/backup_corporativo/settings.py;
+fi
 
 # sobrescrevendo PYTHONPATH
 export PYTHONPATH=$(pwd)/libs
